+++ common/message/codec_header.hpp
@@ -1,6 +1,6 @@
 /***
     This file is part of snapcast
-    Copyright (C) 2014-2022  Johannes Pohl
+    Copyright (C) 2014-2025  Johannes Pohl
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -16,8 +16,8 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ***/
 
-#ifndef MESSAGE_CODEC_HEADER_HPP
-#define MESSAGE_CODEC_HEADER_HPP
+#pragma once
+
 
 // local headers
 #include "message.hpp"
@@ -31,13 +31,15 @@ namespace msg
 class CodecHeader : public BaseMessage
 {
 public:
-    explicit CodecHeader(const std::string& codecName = "", uint32_t size = 0)
-        : BaseMessage(message_type::kCodecHeader), payloadSize(size), payload(nullptr), codec(codecName)
+    /// c'tor taking the @p codec_name and @p site of the payload
+    explicit CodecHeader(const std::string& codec_name = "", uint32_t size = 0)
+        : BaseMessage(message_type::kCodecHeader), payloadSize(size), payload(nullptr), codec(codec_name)
     {
         if (size > 0)
             payload = static_cast<char*>(malloc(size * sizeof(char)));
     }
 
+    /// d'tor
     ~CodecHeader() override
     {
         free(payload);
@@ -54,8 +56,11 @@ public:
         return static_cast<uint32_t>(sizeof(uint32_t) + codec.size() + sizeof(uint32_t) + payloadSize);
     }
 
+    /// payload size
     uint32_t payloadSize;
+    /// the payload
     char* payload;
+    /// name of the codec
     std::string codec;
 
 protected:
@@ -65,7 +70,5 @@ protected:
         writeVal(stream, payload, payloadSize);
     }
 };
+
 } // namespace msg
-
-
-#endif