+++ common/message/hello.hpp
@@ -1,6 +1,6 @@
 /***
     This file is part of snapcast
-    Copyright (C) 2014-2022  Johannes Pohl
+    Copyright (C) 2014-2025  Johannes Pohl
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -16,8 +16,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ***/
 
-#ifndef MESSAGE_HELLO_HPP
-#define MESSAGE_HELLO_HPP
+#pragma once
 
 // local headers
 #include "common/str_compat.hpp"
@@ -31,16 +30,20 @@
 namespace msg
 {
 
+/// Hello message
+/// Initial message, sent from client to server
 class Hello : public JsonMessage
 {
 public:
+    /// c'tor
     Hello() : JsonMessage(message_type::kHello)
     {
     }
 
-    Hello(const std::string& macAddress, const std::string& id, size_t instance) : JsonMessage(message_type::kHello)
+    /// c'tor taking @p macAddress, @p id and @p instance
+    Hello(const std::string& mac_address, const std::string& id, size_t instance) : JsonMessage(message_type::kHello)
     {
-        msg["MAC"] = macAddress;
+        msg["MAC"] = mac_address;
         msg["HostName"] = ::getHostName();
         msg["Version"] = VERSION;
         msg["ClientName"] = "Snapclient";
@@ -51,53 +54,64 @@ public:
         msg["SnapStreamProtocolVersion"] = 2;
     }
 
+    /// d'tor
     ~Hello() override = default;
 
+    /// @return the MAC address
     std::string getMacAddress() const
     {
         return msg["MAC"];
     }
 
+    /// @return the host name
     std::string getHostName() const
     {
         return msg["HostName"];
     }
 
+    /// @return the client version
     std::string getVersion() const
     {
         return msg["Version"];
     }
 
+    /// @return the client name (e.g. "Snapclient")
     std::string getClientName() const
     {
         return msg["ClientName"];
     }
 
+    /// @return the OS name
     std::string getOS() const
     {
         return msg["OS"];
     }
 
+    /// @return the CPU architecture
     std::string getArch() const
     {
         return msg["Arch"];
     }
 
+    /// @return the instance id
     int getInstance() const
     {
         return get("Instance", 1);
     }
 
+    /// @return the protocol version
     int getProtocolVersion() const
     {
         return get("SnapStreamProtocolVersion", 1);
     }
 
+    /// @return a unqiue machine ID
     std::string getId() const
     {
         return get("ID", getMacAddress());
     }
 
+    /// @return a unqiue client ID
     std::string getUniqueId() const
     {
         std::string id = getId();
@@ -109,7 +123,5 @@ public:
         return id;
     }
 };
+
 } // namespace msg
-
-
-#endif