+++ common/message/pcm_chunk.hpp
@@ -38,15 +38,17 @@ namespace msg
 class PcmChunk : public WireChunk
 {
 public:
-    PcmChunk(const SampleFormat& sampleFormat, uint32_t ms)
-        : WireChunk((sampleFormat.rate() * ms / 1000) * sampleFormat.frameSize()), format(sampleFormat), idx_(0)
+    /// c'tor, construct from @p sample_format with duration @p ms
+    PcmChunk(const SampleFormat& sample_format, uint32_t ms) : WireChunk((sample_format.rate() * ms / 1000) * sample_format.frameSize()), format(sample_format)
     {
     }
 
-    PcmChunk() : WireChunk(), idx_(0)
+    /// c'tor
+    PcmChunk() : WireChunk()
     {
     }
 
+    /// d'tor
     ~PcmChunk() override = default;
 
 #if 0
@@ -73,16 +75,18 @@ public:
     //     return result;
     // }
 
-    int readFrames(void* outputBuffer, uint32_t frameCount)
+    /// Read the next @p frame_count frames into @p output_buffer, update the internal read position
+    /// @return number of frames copied to @p output_buffer
+    int readFrames(void* output_buffer, uint32_t frame_count)
     {
         // logd << "read: " << frameCount << ", total: " << (wireChunk->length / format.frameSize()) << ", idx: " << idx;// << "\n";
-        int result = frameCount;
-        if (idx_ + frameCount > (payloadSize / format.frameSize()))
+        int result = frame_count;
+        if (idx_ + frame_count > (payloadSize / format.frameSize()))
             result = (payloadSize / format.frameSize()) - idx_;
 
         // logd << ", from: " << format.frameSize()*idx << ", to: " << format.frameSize()*idx + format.frameSize()*result;
-        if (outputBuffer != nullptr)
-            memcpy(static_cast<char*>(outputBuffer), static_cast<char*>(payload) + format.frameSize() * idx_, format.frameSize() * result);
+        if (output_buffer != nullptr)
+            memcpy(static_cast<char*>(output_buffer), static_cast<char*>(payload) + format.frameSize() * idx_, format.frameSize() * result);
 
         idx_ += result;
         // logd << ", new idx: " << idx << ", result: " << result << ", wireChunk->length: " << wireChunk->length << ", format.frameSize(): " <<
@@ -90,6 +94,8 @@ public:
         return result;
     }
 
+    /// seek @p frames forward or backward
+    /// @return the new read position
     int seek(int frames)
     {
         if ((frames < 0) && (-frames > static_cast<int>(idx_)))
@@ -102,17 +108,20 @@ public:
         return idx_;
     }
 
+    /// @return start time of the current frame
     chronos::time_point_clk start() const override
     {
         return chronos::time_point_clk(chronos::sec(timestamp.sec) + chronos::usec(timestamp.usec) +
                                        chronos::usec(static_cast<chronos::usec::rep>(1000000. * ((double)idx_ / (double)format.rate()))));
     }
 
+    /// @return time of the last frame
     inline chronos::time_point_clk end() const
     {
         return start() + durationLeft<chronos::usec>();
     }
 
+    /// @return duration of this chunk
     template <typename T>
     inline T duration() const
     {
@@ -127,42 +136,51 @@ public:
     //     payloadSize = newSize;
     // }
 
-    void setFrameCount(int frameCount)
+    /// Set the @p frame_count, reserve memory
+    void setFrameCount(int frame_count)
     {
-        auto newSize = format.frameSize() * frameCount;
-        payload = static_cast<char*>(realloc(payload, newSize));
-        payloadSize = newSize;
+        auto new_size = format.frameSize() * frame_count;
+        payload = static_cast<char*>(realloc(payload, new_size));
+        payloadSize = new_size;
     }
 
+    /// @return duration of this chunk in [ms]
     double durationMs() const
     {
         return static_cast<double>(getFrameCount()) / format.msRate();
     }
 
+    /// @return time left, starting from the read pointer
     template <typename T>
     inline T durationLeft() const
     {
         return std::chrono::duration_cast<T>(chronos::nsec(static_cast<chronos::nsec::rep>(1000000 * (getFrameCount() - idx_) / format.msRate())));
     }
 
+    /// @return true if the read pointer is at the end
     inline bool isEndOfChunk() const
     {
         return idx_ >= getFrameCount();
     }
 
+    /// @return number of frames
     inline uint32_t getFrameCount() const
     {
         return (payloadSize / format.frameSize());
     }
 
+    /// @return number of samples
     inline uint32_t getSampleCount() const
     {
         return (payloadSize / format.sampleSize());
     }
 
+    /// Sample format of this chunk
     SampleFormat format;
 
 private:
+    /// current read position (frame idx)
     uint32_t idx_ = 0;
 };
+
 } // namespace msg