+++ changelog.md
@@ -4,8 +4,8 @@
 
 ### Features
 
-- Client: Add support for (secure-) websockets (Issue #1325)
-- Server: Add client authentication (Issue #1334)
+- Client: Add support for (secure-) websockets (Issue #1325, PR #1340)
+- Server: Add client authentication (Issue #1334, PR #1340)
 
 ### Bugfixes